Firestore lemma lookup optimization / Firestore のlemma検索を最適化
